refactor FixedNumberGenerator 클래스를 테스트 util 패키지로 이동
