+++ src/test/java/model/PointGeneratorTest.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.util.FixedNumberGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
